wip: group by assembly district and other geographic regions, almost
